Hi Divya, thanks for covering CSS Regions in your article!
I added some updates to the code samples and to the wording to reflect the latest syntax and approach of the CSS Regions Working Draft, published November 29, 2011.

I also made some changes to the section about region styling. After some debate, the ::region-lines pseudo-element was dropped. A notation similar to media queries is now proposed.